Delete palette
